rename smr (untested)
