+++ sematic/resolver.py
@@ -29,62 +29,6 @@ class Resolver(abc.ABC):
         """
         pass
 
-    @abc.abstractclassmethod
-    def activate_resource_for_run(
-        cls, resource: AbstractExternalResource, run_id: str, root_id: str
-    ) -> AbstractExternalResource:
-        """Associate the provided resource with the given run and activate it.
-
-        This call will block until the resource is in either the ACTIVE state
-        or a terminal state (in the case of failure to activate the resource).
-        If the resource can't be activated, will raise ExternalResourceError.
-
-        Parameters
-        ----------
-        resource:
-            The external resource to activate
-        run_id:
-            The id of the future/run that is using the resource
-        root_id:
-            The id of the root run for the given run_id
-
-        Returns
-        -------
-        The active version of the resource
-
-        Raises
-        ------
-        ExternalResourceError:
-            If the external resource could not be activated
-        """
-        pass
-
-    @abc.abstractclassmethod
-    def deactivate_resource(cls, resource_id: str) -> AbstractExternalResource:
-        """Deactivate the resource with the given id.
-
-        This call should block until the resource is deactivated or the
-        resource has failed to deactivate. If the resource fails to deactivate,
-        will raise an ExternalResourceError.
-
-        This may be called even if the resource is already deactivated.
-
-        Parameters
-        ----------
-        resource_id:
-            The id of the resource to deactivate
-
-        Returns
-        -------
-        The deactivated resource.
-
-        Raises
-        ------
-        ExternalResourceError:
-            If the resource fails to deactivate.
-        """
-        pass
-
     @classmethod
     def entering_resource_context(cls, resource: AbstractExternalResource):
         """A hook resolvers may use to take action once a resource is activated.
